pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "caasl"
dynamic = ["version"]
description = "Causal Amortized Structure Learning with Deep Reinforcement Learning"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Yashas Annadani", email = "yashas.annadani@example.com"}
]
maintainers = [
    {name = "Yashas Annadani", email = "yashas.annadani@example.com"}
]
keywords = ["causal-inference", "reinforcement-learning", "deep-learning", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.14.5",
    "torch>=1.9.0",
    "torchvision",
    "gym>=0.21.0",
    "wandb",
    "igraph",
    "scipy",
    "setuptools",
    "dowel",
    "avici",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
    "pre-commit",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
]

[project.urls]
Homepage = "https://github.com/your-username/caasl"
Documentation = "https://github.com/your-username/caasl#readme"
Repository = "https://github.com/your-username/caasl.git"
"Bug Tracker" = "https://github.com/your-username/caasl/issues"

[project.scripts]
caasl = "caasl.cli:main"

[tool.setuptools_scm]
write_to = "caasl/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["caasl"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=caasl",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["caasl"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
] 